bash-completion: get shells from docker image
The list of shells is now completed using the file /etc/shells read from
the docker image.

The environment variable COMPLETION_NODOSH can prevent from running dosh
command to auto complete the command-line.

If set, the list of shells is not read from /etc/shells. The completion
for --shell falls back to the list of local files.