tests.sh: fix some comparaison tests
Those tests are wrongs. They cannot rely on the stderr output of the
default shell interpreter (aka /bin/sh) as the implementation (and the
version) in use may differ between the host and the container. Debian
uses dash and Archlinux uses bash. As a consequence, the tests should
rely on the exit status returned by the interpreter.