Fix and simplify the handling of option -c
This commit fixes an issue introduced with commit 92da05e (Fix name and
argument issue with -c) in the case of the command line sets the option
-c without giving a command.

In that situation, the shift fails that causes dosh to exit without
running the docker command. This case should be handle by the shell
interpreter that is run by docker-run; dosh is a wrapper only.

Note: This case would have causes some tests to fail before the commit
5cd1cf7 (tests.sh: fix some comparaison tests) have fixed them.

This commit terminates the simplification of the handling of the option
-c as it was started by commit 816891a (Handle -c, -i and -s shell
options in is_shopt()). As it was said above, dosh is a wrapper script
only, as a consequence, it does not distinguish anymore the options -c,
from the options -i, -s.

As a final word, that commit prepends the array of arguments "$@" with
the shopts only if it is not empty.

[gportay: cherry-pick with adaptations (fb3790d)]