Handle shopts optional argument
Optional arguments are not handled. The shell option -|+O takes an
optional argument.

The command below opens a new shell and is valid:

	$ bash +O

This fixes the behavior of the following command that fails quietly
after shifting the "missing" argument (due to set -e and no arguments
left):

	$ dosh +O

The trace of the failure:

	$ bash -x dosh -O
	(...)
	+ is_shopt_argument -O
	+ [[ -O =~ ^[-+][o]$ ]]
	+ '[' bash == bash ']'
	+ [[ -O =~ ^[-+]O$ ]]
	+ return 0
	+ shopts+=("$1" "$2")
	+ eval 'opt_O=""'
	++ opt_O=
	+ shift
	+ shift <- fails because there is no arguments left