Use the content of the DOCKER variable in an array
The environment variable DOCKER may contains more that one word, as in
DOCKER="sudo docker".

This is why the docker variable was not double quoted; but it does not
prevent from the globbing and word splitting.

This commits uses the content of that variable as a bash array to fixes
the shellcheck warning SC2086.

Note: The comments to ignore that shellcheck warning in the functions
docker_{build|rmi}() cannot be removed because the variables
DOSH_DOCKER_{BUILD|RMI}_EXTRA_OPTS are declared after the definition of
those functions.

[1]: https://github.com/koalaman/shellcheck/wiki/SC2086