Use the content of docker variables in an array
The comments to ignore that shellcheck warning in the functions
docker_{build|rmi}() cannot be removed because the variables
DOSH_DOCKER_{BUILD|RMI}_EXTRA_OPTS are declared after the definition of
those functions.

This commits uses the content of those variable as a bash array and
append it to another array to fixes the shellcheck warning SC2086.

[1]: https://github.com/koalaman/shellcheck/wiki/SC2086