Use the content of the first parameter in an array
If dosh is used as a shebang and if it has arguments, then bash gives
all the arguments in the very first parameter only (i.e. $1).

The given shebang:

	#!/usr/bin/dosh --dockerfile Dockerfile.alpine -x

Gives "--dockerfile Dockerfile.alpine -x" as first parameter (i.e. $1).

This is why the args variable was not double quoted; but it does not
prevent from the globbing and word splitting.

This commits uses the content of that variable as a bash array to fixes
the shellcheck warning SC2086.