Add missing final dot to some options
Both commits e5f42f9 (Add support for user-wide ~/.dosh_profile) and
f0fee38 (Add --no-doshrc option) have introduced two options.

The descriptions of those options in the usage and the man-page do not
have the final dot; the other options have.

This commit adds the missing final dot to the description of the options
--no-doshrc and --no-doshprofile.