Fix retrieval of the docker group-id
The commit 5937c41 (Support for dosh-in-dosh) has introduced support for
dosh-in-dosh. It retrieves the docker group-id to make the current user
member of it in the container.

The gid is retrieved using the grep expresion '^docker' that may lead to
a misgroup-id issue if a group exists with a name that starts with
docker (i.e. dockerwhatever...).

This commit fixes the retrieval of the docker gid by adding the
separator character (i.e. colon) to the end of the group name.