Fallback to user group if docker group does not exist
The commit 5937c41 (Support for dosh-in-dosh) has introduced support for
dosh-in-dosh. It retrieves the docker group-id to make the current user
member of it in the container.

The variable did is left empty if the docker group does not exist; that
unbound variable causes dosh to exit failure since commit 53f2463 (Uses
the unset flag).

This commit replaces the here-string redirection (<<<) and its command
substitution ($()) by the input redirection (<) and its process
substition (<()). This lets read fail if nothing is read (i.e. the group
docker does not exist). The commits falls back to the user group then to
make sure the variable did is bound.

Fixes:

	dosh: line 155: did[2]: unbound variable