Fix missing blank line in zsh options
The zsh options were added by the commit 4230e2a (Add support for zsh
options).

The list of the zsh options are displayed on the same line as the text
saying `Zsh specific options' because there is no blank line between
both.

This commit adds the missing blank line after `Zsh specific options' and
the list of the zsh options.

Fixes:

	Zsh specific options: -0123456789dgklprtwyBDEFGHIJKLMNOPQRSTUVWXYZ, +0123456789dgklprtwyBDEFGHIJKLMNOPQRSTUVWXYZ
		For a more thorough description of shopts, please refers to zsh help.