Fix forwarding -e extra docker run or exec option
The docker short option -e clashes with the command echo used to convert
the DOSH_DOCKER_RUN_EXTRA_OPTS and DOSH_DOCKER_EXEC_EXTRA_OPTS variables
to an array.

It is interpreted as an option given to echo if it is given as first
argument to DOSH_DOCKER_RUN_EXTRA_OPTS or DOSH_DOCKER_EXEC_EXTRA_OPTS
variable. Therefore, it is not added as an additionnal parameter to the
docker run or exec command.

This fixes the forwarding of the extra option -e to docker run and exec
commands by using the here strings redirection instead of command echo.

According to man bash(1):

	Here Strings

		A variant of here documents, the format is:

		[n]<<<word

	The word undergoes tilde expansion, parameter and variable
	expansion, command substitution, arithmetic expansion, and quote
	removal. Pathname expansion and word splitting are not
	performed. The result is supplied as a single string, with a
	newline appended, to the command on its standard input (or file
	descriptor n if n is specified).

	(...)

	echo [-neE] [arg ...]

	Output the args, separated by spaces, followed by a newline. The
	return status is 0 unless a write error occurs. If -n is
	specified, the trailing newline is suppressed. If the -e option
	is given, interpretation of the following backslash-escaped
	characters is enabled. The -E option disables the interpretation
	of these escape characters, even on systems where they are
	interpreted by default. The xpg_echo shell option may be used to
	dynamically determine whether or not echo expands these escape
	characters by default. echo does not interpret -to mean the end
	of options.

According to man echo(1):

	-e     enable interpretation of backslash escapes

According to man docker-run(1):

	-e, --env=[]

		Set environment variables

	This option allows you to specify arbitrary environment
	variables that are available for the process that will be
	launched inside of the container.