Fix forwarding extra docker options with space
The use of read -r -a and here strings do not preserve the spaces as it
is used to split the given here string into an array of words (unless
the spaces are escaped by a backslash in the here string and the read
option -r is unset; but the -r option is mantained for now).

Unfortunatly, there is no simple solution in bash to maintain set -u and
to do testing for array variables that are unbound. It is even worse if
targeting the compatibily accross all the verions of bash.

The most simple solution is to modify the algorithmic of the bash script
in order to consider the docker extra options as an array.

This implies to convert the variables given through the environment from
a string to an array using the command read -a. That conversion is
performed before the source scripts are read and executed.

Important: This introduces a breaking change with the previous versions
as the profile files **MUST** use bash arrays for the docker variables
and they **CANNOT** use strings anymore as they are not splitted into
words.

According to man read(1p):

	-r        Do not treat a <backslash> character in any special
		  way. Consider each <backslash> to be part of the input
	          line.

According to man bash(1):

	echo [-neE] [arg ...]

	Output the args, separated by spaces, followed by a newline. The
	return status is 0 unless a write error occurs. If -n is
	specified, the trailing newline is suppressed. If the -e option
	is given, interpretation of the following backslash-escaped
	characters is enabled. The -E option disables the interpretation
	of these escape characters, even on systems where they are
	interpreted by default. The xpg_echo shell option may be used to
	dynamically determine whether or not echo expands these escape
	characters by default. echo does not interpret -to mean the end
	of options.

	read [-ers] [-a aname] [-d delim] [-i text] [-n nchars] [-N nchars] [-p prompt] [-t timeout] [-u fd] [name ...]

	One line is read from the standard input, or from the file
	descriptor fd supplied as an argument to the -u option, split
	into words as described above under Word Splitting, and the
	first word is assigned to the first name, the second word to the
	second name, and so on. If there are more words than names, the
	remaining words and their intervening delimiters are assigned to
	the last name. If there are fewer words read from the input
	stream than names, the remaining names are assigned empty
	values. The characters in IFS are used to split the line into
	words using the same rules the shell uses for expansion
	(described above under Word Splitting). The backslash character
	(\) may be used to remove any special meaning for the next
	character read and for line continuation.  Options, if supplied,
	have the following meanings:

		-a aname

			The words are assigned to sequential indices of
			the array variable aname, starting at 0. aname
			is unset before any new values are assigned.
			Other name arguments are ignored.

		-r	Backslash does not act as an escape character.
			The backslash is considered to be part of the
			line. In particular, a backslash-newline pair
			may not then be used as a line continuation.

	(...)

	-u      Treat unset variables and parameters other than the
		special parameters "@" and "*" as an error when
		performing parameter expansion. If expansion is
		attempted on an unset variable or parameter, the shell
		prints an error message, and, if not interactive, exits
	        with a non-zero status.