Fix interactive shell if DOSHELL is unset and non posix option is given
The parsing of the argument breaks after the first non dosh option is
found. It breaks if a non $DOSHELL option is found as well, it implies
$DOSHELL to be set to a specific shell otherwise (bash, dash, zsh...).

The variable $DOSHELL defaults to /bin/sh and both functions is_shopt()
and is_shopt_argument() do not check for the shell specific options in
order to continue the parsing of the arguments.

Both docker-run options pseudo-TTY and interactive are unset as there
are remaining positional parameters at that point if a specific shell
option is given (unless -i or -s are given).

It is expected the command dosh --shell /bin/bash -l runs bash -l in a
container with an allocated pseudo-TTY and STDIN opened.

But the container exits immediately because no pseudo-TTY is allocated
and the container is not interactive:

	host$ dosh -l
	host$ echo $?
	0

This fixes the parsing of the non posix specific options by checking for
the given option against the shell set by the dosh option --shell or the
dosh shell set in $DOSHELL if unset.

Note: The container never exits if a pseudo-TTY is not allocated or
STDIN is not opened. Neither Ctrl-Z, Ctrl-C or Ctrl-\ kills the process,
and the container has to be stopped using docker kill:

	$ dosh -l
	(...)
	^Z
	(...)
	^C
	(...)
	^\
	(...)