Fix interactive shell if DOSHELL set in profile and non posix option is given
See commit ab1226df5a8cdc65f6dc3449cc27036fc8badb99.

The variable $DOSHELL defaults to /bin/sh and it can be set via the dosh
option --shell or via the profile files ~/.dosh_profile or ./doshrc.

It is expected the command dosh -l runs bash -l in a container with an
allocated pseudo-TTY and STDIN opened if the $DOSHELL is set to bash
after the parsing of arguments (i.e. via a dosh profile files).

But the container exits immediately because no pseudo-TTY is allocated
and the container is not interfactive as the option -l is a non posix
and $DOSHELL is set to /usr/bin during the parsing of the arguments.

This fixes the parsing of the non posix specific options if $DOSHELL is
set afterward in the profile files by checking for positional arguments
after the arguments are actually parsed.