dosh: fix extra arguments containing whitespaces
The environment variables DOSH_DOCKER_*_EXTRA_OPTS are shell strings
representing a list of command line arguments separated by whitespaces.

They are strings as if they are given to /bin/sh -c "command_string", at
the exception of the command_string is a stringified list of arguments
of a very single command.

The strings are splitted to a bash array using the whitespace character
as delimiter. Every single space is taken into account without any
distinction.

However, there are situtations where the whitespace is a part of the
argument itself (i.e. it is not a delimiter).

For example:

	DOSH_DOCKER_RUN_EXTRA_OPTS="--env 'FOO=bar baz'"

In that example, the string is splitted into the three arguments:
 - --env
 - 'FOO=bar
 - baz'

But the second whitespace is not intended to get used for splitting by
the read built-in.

This removes the read option -r to let a way to escape characters using
a backslash, and thus to prevent the read built-in to split a whitespace
preceded by a backslash.

The example above have to be rewritten this way:

	DOSH_DOCKER_RUN_EXTRA_OPTS="--env FOO=bar\ baz"

Note: It is the more trivial change, even though is not perfect.