cqfd: restore IFS after call to cfg_parser()
The function cfg_parser modifies (i.e. sets) the variable IFS[1] and
does not restore it.

As a consequence, the function modifies the word splitting[2] for the
entire script after it is called. The "words" are split by lines and
"breaks" the default word splitting by whitespaces.

This makes the variable IFS local to the function so the previous value
is automatically "restored" once the function exits.

Fixes:

	gportay@archlinux ~/src/dosh $ support/cqfd release
	+ set -e
	+ set -o pipefail
	+ asciidoctor -b manpage -o - dosh.1.adoc
	+ gzip -c -
	tar: dosh dosh.1.gz: Cannot stat: No such file or directory
	xz: Reduced the number of threads from 12 to 3 to not exceed the memory usage limit of 3,926 MiB
	tar: Exiting with failure status due to previous errors

[1]: https://www.gnu.org/software/bash/manual/bash.html#index-IFS
[2]: https://www.gnu.org/software/bash/manual/html_node/Word-Splitting.html