cqfd: fix export of DOSH_ variables if set in .cqfdrc
The stolen function cfg_parser()[1] substitutes the ini properties into
bash arrays.

As of bash(1):

	BUGS

	(...)

	Array variables may not (yet) be exported.

Therefore, the DOSH_ variables set in the .cqfdrc cannot be exported as
is; the arrays have to be unset before being exported as environment
variables.

This unsets the DOSH_ arrays before exporting them as environment
variables.

Tested:

	$ support/cqfd -b podman run env
	HOSTNAME=5859b55642b2
	HOME=/home/gportay
	container=podman
	TERM=xterm
	PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
	SSH_AUTH_SOCK=/home/gportay/.sockets/ssh
	DOSHLVL=1

[1]: https://ajdiaz.wordpress.com/2008/02/09/bash-ini-parser/