Makefile: add several shell targets
This adds several shell targets for the most used shells.

Note: The target runs dosh -c "$SHELL", i.e. it does not run the $SHELL
directly. As a consequence, dosh does not go in interactive mode and
does not allocate a pseudo-TTY as not interactive flag is explicitly set
and the shell command has arguments. Thus, the interactive $SHELL ends.
The targets appends explicitly the interactive option via the variable
.SHELLFLAGS[1].

[1]: https://www.gnu.org/software/make/manual/html_node/Choosing-the-Shell.html