dosh: list: fix tag column alignment for if platform images
This fixes commit a5b45face65ee34afb21803b429a95f683355054.

The images do not have the same lenght since platform was introduced.

This fixes the column alignment if image has platform. It iterates the
whole list of images and takes the image with the longest tag to set the
column alignment.